Unified ```task_type``` resolving due to https://github.com/mrgambal/vulyk/commit/8ad16ddd856437629a766b3187f682a4365a7481
